requirements.txt
Django==4.2.21
Pillow==11.2.1
python-slugify==8.0.4
django-ckeditor==6.7.0
django-taggit==5.0.1
gunicorn==23.0.0
dotenv==0.9.9
